query BDA 40:96 before every INT16 call, as it may change
It happens that 40:96 may be zero until an extended key was pressed.
This made EDIT fail to detect an extended keyboard properly on some
machines, and the keys accessed via AltGr were not recognized, because
AltGr is mapped to ALT when using INT16.02. All ALT combinations
however are routed to the menu handler.